backend/requirements.txt
google-generativeai
opencv-python
fer
faiss-cpu
sentence-transformers
python-dotenv
numpy
tensorflow
flask
flask-cors
flask-jwt-extended
